django-backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.7
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-import-export==4.3.7
django-timezone-field==7.1
djangorestframework==3.16.0
gunicorn==23.0.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fonttools==4.56.0
fpdf2==2.8.2
idna==3.10
inflection==0.5.1
kombu==5.5.3
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.51
psutil==7.0.0
PyJWT==2.9.0
PyPDF2==3.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
opencv-python==4.11.0.86
opencv-python-headless ==4.11.0.86
numpy==2.2.6
torch==2.7.1
torchvision==0.22.1
django-prometheus==2.4.1
anomalib==2.0.0
lightning==2.5.1.post0
kornia==0.8.1
tifffile==2025.5.10
matplotlib==3.10.3
pandas==2.2.3
scikit-learn==1.7.0
scikit-image==0.25.2
openvino==2025.1.0
timm==1.0.15
einops==0.8.1
freia==0.2
open-clip-torch==2.24.0